Fix some cross-wordsize compilation printf() format warnings
We remove many uint64_t references, which seem to lack a sensible
way to be represented on architectures with different wordsizes.

Since we also prefer not to use the (non-portable) %zu format char,
which is mainly used for sizeof() printf()'s, we turn sizeof() into
a macro and cast it to an unsigned int, which is portably printable
as %u.

+++ hooks.c
@@ -16,10 +16,10 @@
 
 static int check_dupes;
 static merlin_event last_pkt;
-static uint64_t dupes, dupe_bytes;
+static unsigned long long dupes, dupe_bytes;
 
 struct check_stats {
-	uint64_t poller, peer, self;
+	unsigned long long poller, peer, self;
 };
 static struct check_stats service_checks, host_checks;
